pyproject.toml
[project]
name = "rfmeter"
version = "0.1.0"
description = "ImmersionRC RF Power Meter logger"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "pandas>=2.3.1",
    "pyqt6>=6.9.1",
    "pyserial>=3.5",
    "typer>=0.15",
]

[project.scripts]
rfmeter = "rfmeter.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
